funcion cli
